+++ src/window.cpp
@@ -9,7 +9,7 @@
 #include <chrono>
 #include <thread>
 
-Window::Window(const std::string& title, int width, int height, bool handles_updates, int target_delta)
+Window::Window(const std::string& title, int width, int height, bool handles_updates, float target_delta)
     : window(sf::VideoMode(width, height), title, sf::Style::Titlebar | sf::Style::Resize | sf::Style::Close),
     target_delta(target_delta), handles_updates(handles_updates) {}
